backend/requirements.txt
# =====================================
# MRD SYSTEM - REQUIREMENTS
# Medical Records Database System
# =====================================

# Django Framework & Core
Django>=5.2.6
djangorestframework>=3.15.0

# Database Support
psycopg2-binary>=2.9.0

# Authentication & Security
django-cors-headers>=4.4.0
django-ratelimit>=4.1.0

# API Enhancement
django-filter>=24.0
django-extensions>=3.2.0
drf-spectacular>=0.27.0

# Background Tasks & Caching
celery>=5.3.0
redis>=5.0.0

# Configuration Management
python-decouple>=3.8
python-dotenv>=1.0.0

# File Processing & Documents
Pillow>=10.0.0
reportlab>=4.0.0
openpyxl>=3.1.0
pandas>=2.0.0
xlsxwriter>=3.1.0

# Date/Time Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# HTTP & API Utilities
requests>=2.31.0

# Development & Debugging
django-debug-toolbar>=4.2.0

# Testing
pytest>=7.4.0
pytest-django>=4.8.0

# Production Utilities
gunicorn>=21.2.0
whitenoise>=6.6.0
twilio>=8.0.0

# Error Tracking & Monitoring
sentry-sdk>=2.0.0
python-json-logger>=2.0.7
